Move the storage module and upload_worker to the utils pakage in order to run it without hub
