+++ biothings/__init__.py
@@ -1,5 +1,8 @@
 from typing import NamedTuple
 
+# stub, until set to real config module
+config = None
+
 
 class _version_info(NamedTuple):
     # similar to sys._version_info
@@ -14,7 +17,11 @@ __version__ = '.'.join(map(str, version_info))
 
 def config_for_app(config):
     import warnings
-    warnings.warn((
-        "It is safe to remove this function call now. "
-        "import biothings.hub will take care of it."
-    ), DeprecationWarning)
+
+    warnings.warn(
+        (
+            "It is safe to remove this function call now. "
+            "import biothings.hub will take care of it."
+        ),
+        DeprecationWarning,
+    )
