+++ biothings/hub/dataload/uploader.py
@@ -2,7 +2,6 @@ import asyncio
 import copy
 import datetime
 import inspect
-import logging as loggingmod
 import os
 import time
 from functools import partial
@@ -14,8 +13,8 @@ from biothings.utils.hub_db import get_src_dump, get_src_master, get_src_conn
 from biothings.utils.loggers import get_logger
 from biothings.utils.manager import BaseSourceManager, ResourceNotFound
 from biothings.utils.version import get_source_code_info
-
-from .storage import (
+from biothings.utils.workers import upload_worker
+from biothings.utils.storage import (
     BasicStorage,
     IgnoreDuplicatedStorage,
     MergerStorage,
@@ -32,55 +31,6 @@ class ResourceNotReady(Exception):
 
 class ResourceError(Exception):
     pass
-
-
-def upload_worker(name, storage_class, loaddata_func, col_name, batch_size, batch_num, *args, **kwargs):
-    """
-    Pickable job launcher, typically running from multiprocessing.
-    storage_class will instanciate with col_name, the destination
-    collection name. loaddata_func is the parsing/loading function,
-    called with `*args`.
-    """
-    data = []
-    db = None
-    if 'db' in kwargs:
-        db = kwargs.get('db')
-    try:
-        data = loaddata_func(*args)
-        if type(storage_class) is tuple:
-            klass_name = "_".join([k.__class__.__name__ for k in storage_class])
-            storage = type(klass_name, storage_class, {})(None, col_name, loggingmod)
-        else:
-            storage = storage_class(db, col_name, loggingmod)
-        return storage.process(data, batch_size)
-    except Exception as e:
-        logger_name = "%s_batch_%s" % (name, batch_num)
-        logger, logfile = get_logger(logger_name, config.LOG_FOLDER)
-        logger.exception(e)
-        logger.error(
-            "Parameters:\nname=%s\nstorage_class=%s\n" % (name, storage_class)
-            + "loaddata_func=%s\ncol_name=%s\nbatch_size=%s\n"
-            % (loaddata_func, col_name, batch_size)
-            + "args=%s" % repr(args)
-        )
-        import pickle
-
-        pickfile = os.path.join(os.path.dirname(logfile), "%s.pick" % logger_name)
-        try:
-            pickle.dump(
-                {
-                    "exc": e,
-                    "params": {"name": name, "storage_class": storage_class},
-                    "loaddata_func": loaddata_func,
-                    "col_name": col_name,
-                    "batch_size": batch_size,
-                    "args": args,
-                },
-                open(pickfile, "wb"),
-            )
-        except TypeError as ie:
-            logger.warning("Could not pickle batch errors: %s" % ie)
-        raise e
 
 
 class BaseSourceUploader(object):
